+++ cnccode/business_classes/BURenewal.inc.php
@@ -383,6 +383,8 @@ class BURenewal extends Business
 
             $row['salePrice'] = null;
             $row['costPrice'] = null;
+            $row['units'] = $dbeJRenContract->getValue(DBEJRenContract::users);
+            $row['directDebit'] = $dbeJRenContract->getValue(DBEJRenContract::directDebitFlag) == 'Y';
             if ($displayAccountsInfo) {
                 $row['salePrice'] = $dbeJRenContract->getValue(DBEJRenContract::curUnitSale);
                 $row['costPrice'] = $dbeJRenContract->getValue(DBEJRenContract::curUnitCost);
@@ -456,6 +458,8 @@ class BURenewal extends Business
                 );
             $row['salePrice'] = null;
             $row['costPrice'] = null;
+            $row['units'] = $dbeJRenContract->getValue(DBEJRenContract::users);
+            $row['directDebit'] = $dbeJRenContract->getValue(DBEJRenContract::directDebitFlag) == 'Y';
             if ($displayAccountsInfo) {
                 $row['salePrice'] = $dbeJRenDomain->getValue(DBEJRenDomain::salePrice);
                 $row['costPrice'] = $dbeJRenDomain->getValue(DBEJRenDomain::costPrice);
@@ -511,7 +515,8 @@ class BURenewal extends Business
                 $row['salePrice'] = $dbeJRenBroadband->getValue(DBEJRenBroadband::salePricePerMonth) * 12;
                 $row['costPrice'] = $dbeJRenBroadband->getValue(DBEJRenBroadband::costPricePerMonth) * 12;
             }
-
+            $row['units'] = $dbeJRenContract->getValue(DBEJRenContract::users);
+            $row['directDebit'] = $dbeJRenContract->getValue(DBEJRenContract::directDebitFlag) == 'Y';
             $row['description'] = $dbeJRenBroadband->getValue(DBEJRenBroadband::itemDescription);
             $row['customerItemID'] = $dbeJRenBroadband->getValue(DBEJRenBroadband::customerItemID);
             $row['itemTypeDescription'] = $dbeJRenBroadband->getValue(DBEJRenBroadband::itemTypeDescription);
@@ -560,7 +565,8 @@ class BURenewal extends Business
                 $row['salePrice'] = $dbeJRenHosting->getValue(DBEJRenHosting::curUnitSale);
                 $row['costPrice'] = $dbeJRenHosting->getValue(DBEJRenHosting::curUnitCost);
             }
-
+            $row['units'] = $dbeJRenContract->getValue(DBEJRenContract::users);
+            $row['directDebit'] = $dbeJRenContract->getValue(DBEJRenContract::directDebitFlag) == 'Y';
             $row['description'] = $dbeJRenHosting->getValue(DBEJRenHosting::itemDescription);
             $row['customerItemID'] = $dbeJRenHosting->getValue(DBEJRenHosting::customerItemID);
             $row['itemTypeDescription'] = $dbeJRenHosting->getValue(DBEJRenHosting::itemTypeDescription);
@@ -620,6 +626,8 @@ class BURenewal extends Business
             $row['itemTypeDescription'] = $dbeJRenQuotation->getValue(DBEJRenQuotation::itemTypeDescription);
             $row['notes'] = $dbeJRenQuotation->getValue(DBEJRenQuotation::notes);
             $row['expiryDate'] = $dbeJRenQuotation->getValue(DBEJRenQuotation::nextPeriodStartDate);
+            $row['units'] = $dbeJRenContract->getValue(DBEJRenContract::users);
+            $row['directDebit'] = $dbeJRenContract->getValue(DBEJRenContract::directDebitFlag) == 'Y';
             $row['renewalTypeID'] = 3;
             $row['coveredItems'] = [];
             $row['calculatedExpiryDate'] = (